Update MockConsumer.java
The `addRecord` implementation can be very inefficient given a large number of topics, partitions, and records.  The problem is that entire set of `TopicPartitions` is copied every time `addRecord` is invoked in order to check if a record is in the consumer's assigned set.  Instead, if we can just query if a `TopicPartition` is in the assignment in O(1) time.

Tested this with 100 topics, 40 partitions, and 800 added records.  Prior to the change, the test took at least a minute to run (didn't actually wait for it to finish).  After the change, the test ran in under a second.